Rename municipality to city
